requirements.txt
Flask>=0.10
Flask-WTF==0.10.2
Flask-SQLAlchemy==2.0
NumPy>=1.7.1
SciPy>=0.17
scikit-learn>=0.17